♻️ Move step_logger.go to internal/wrap
