+++ internal/wrap/opts_aliases.go
@@ -1,0 +1,5 @@
+package wrap
+
+import "github.com/akm/sql-slog/opts"
+
+type StepOptions = opts.StepOptions
